Rename to optimization_steps and let it able to run from any dir
