+++ src/test/CMakeLists.txt
@@ -10,8 +10,7 @@ file(GLOB_RECURSE TEST_CONTEXT_SHIFT "test_context_shift.cpp")
 file(GLOB_RECURSE TEST_EQUIV_SET "test_equivalence_set.cpp")
 file(GLOB_RECURSE TEST_GENERATOR "test_generator.cpp")
 file(GLOB_RECURSE TEST_GRAPH_TO_QASM "test_graph_to_qasm.cpp")
-file(GLOB_RECURSE TEST_OPTIMIZATION "test_optimization.cpp")
-file(GLOB_RECURSE TEST_OPTIMIZATION_SA "test_optimization_sa.cpp")
+file(GLOB_RECURSE TEST_OPTIMIZATION_STEPS "test_optimization_steps.cpp")
 file(GLOB_RECURSE TEST_PHASE_SHIFT "test_phase_shift.cpp")
 file(GLOB_RECURSE TEST_PRUNING "test_pruning.cpp")
 file(GLOB_RECURSE TEST_QUARTZ "test_quartz.cpp")
@@ -46,8 +45,7 @@ add_executable(test_context_shift ${TEST_CONTEXT_SHIFT} )
 add_executable(test_equiv_set ${TEST_EQUIV_SET} )
 add_executable(test_generator ${TEST_GENERATOR} )
 add_executable(test_graph_to_qasm ${TEST_GRAPH_TO_QASM} )
-add_executable(test_optimization ${TEST_OPTIMIZATION} )
-add_executable(test_optimization_sa ${TEST_OPTIMIZATION_SA} )
+add_executable(test_optimization_steps ${TEST_OPTIMIZATION_STEPS} )
 add_executable(test_phase_shift ${TEST_PHASE_SHIFT} )
 add_executable(test_pruning ${TEST_PRUNING} )
 add_executable(test_quartz ${TEST_QUARTZ} )
